+++ cellprofiler_core/image/abstract_image_provider/_abstract_image_provider.py
