+++ tests/modules/test_namesandtypes.py
@@ -16,6 +16,7 @@ import cellprofiler_core.modules.namesandtypes
 import cellprofiler_core.modules.loadimages
 import cellprofiler_core.object
 import cellprofiler_core.pipeline
+import cellprofiler_core.utilities.pathname
 import cellprofiler_core.workspace
 import tests.modules
 
@@ -1451,7 +1452,7 @@ def run_workspace(
     n.module_num = 1
     pipeline = cellprofiler_core.pipeline.Pipeline()
     pipeline.add_module(n)
-    url = cellprofiler_core.modules.loadimages.pathname2url(path)
+    url = cellprofiler_core.utilities.pathname.pathname2url(path)
     pathname, filename = os.path.split(path)
     m = cellprofiler_core.measurement.Measurements()
     if load_as_type == cellprofiler_core.modules.namesandtypes.LOAD_AS_OBJECTS:
@@ -1531,7 +1532,7 @@ def run_workspace(
         si.rescale.value = rescaled
         si.manual_rescale.value = manual_rescale
 
-        url = cellprofiler_core.modules.loadimages.pathname2url(path)
+        url = cellprofiler_core.utilities.pathname.pathname2url(path)
         pathname, filename = os.path.split(path)
         if load_as_type == cellprofiler_core.modules.namesandtypes.LOAD_AS_OBJECTS:
             url_feature = cellprofiler_core.measurement.C_OBJECTS_URL + "_" + name
